overhaul: negation, apply wrt normalized SDD
Negation: got rid of complemented edges since they complicated things
and were a constant source of bugs. Negation is therefore computed
explicitly and cached alongside the node.

Apply: removed assumptions about apply w.r.t. normalized SDD and
followed the referential implementation.

+++ sdd-rs-lib/dot_writer.rs
@@ -30,7 +30,7 @@ pub struct DotWriter {
 
 pub enum NodeType {
     Box(String),
-    Circle(u32),
+    Circle(u16),
     CircleStr(String),
     Record(String, String),
 }
